requirements.txt
fastapi
uvicorn
torch
transformers
Pillow
numpy
python-multipart